+++ index.php
@@ -19,7 +19,7 @@ require_once __DIR__ . '/config.php';
 // If we detect that an OPDS reader try to connect try to redirect to feed.php
 if (preg_match('/(Librera|MantanoReader|FBReader|Stanza|Marvin|Aldiko|Moon\+ Reader|Chunky|AlReader|EBookDroid|BookReader|CoolReader|PageTurner|books\.ebook\.pdf\.reader|com\.hiwapps\.ebookreader|OpenBook)/', $_SERVER['HTTP_USER_AGENT'])) {
     header('location: ' . Config::ENDPOINT["feed"]);
-    exit();
+    return;
 }
 
 $request = new Request();